Fixed migration on reset, removed matchesTime, simplified compare, added DataManager properties

+++ app/database/migrations/2015_09_29_154442_change_old_differentiated_widgets.php
@@ -61,7 +61,9 @@ class ChangeOldDifferentiatedWidgets extends Migration {
         }
         /* Deleting descriptors. */
         foreach ($deletableDescriptors as $descriptor) {
-            $descriptor->delete();
+            if ($descriptor) {
+                $descriptor->delete();
+            }
         }
     }
